Renames useFormInputsRef
