+++ src/useFormRef/useFormRef.ts
@@ -1,6 +1,6 @@
 import { Ref } from 'react';
 import { ReactOrNativeEventListener } from '../useSettersAsEventHandler';
-import { useAnchorInputs } from './useAnchorInputs';
+import { useAnchorInputs } from './useFormInputsRef';
 import { useFormAnchor } from './useFormAnchor';
 import { useHandlers } from '../useHandlers';
 import { useSettersAsRefEventHandler } from '../useSettersAsRefEventHandler';
